+++ editors/phpstorm/src/main/kotlin/com/atatusoft/ppphp/PpphpLanguage.kt
@@ -1,0 +1,11 @@
+package com.atatusoft.ppphp
+
+import com.intellij.lang.Language
+import com.jetbrains.php.lang.PhpLanguage
+
+class PpphpLanguage private constructor() : Language(PhpLanguage.INSTANCE, "++PHP") {
+    companion object {
+        @JvmField
+        val INSTANCE = PpphpLanguage()
+    }
+}